chore(FTDA): Update folder structure & repositioned useEffect hook
