+++ src/Pages/HomePage/components/CreateTask/CreateTask.tsx
@@ -1,13 +1,14 @@
+import { useForm, SubmitHandler, Controller } from "react-hook-form";
 import { Button, Modal, Select, TextInput, Textarea } from "@mantine/core";
 import { DateInput } from "@mantine/dates";
-import { useForm, SubmitHandler, Controller } from "react-hook-form";
-import { Task } from "../../types/Task";
+
+import { Task } from "../../../../types/Task";
 import "@mantine/dates/styles.css";
 import {
   addTaskToLocalStorage,
   getLastTaskId,
   saveLastTaskId,
-} from "../../localstorage/localstorage";
+} from "../../../../localstorage/localstorage";
 
 type CreateTaskProps = {
   opened: boolean;
